chore: Refactor client and server
