+++ packages/client/global.d.ts
@@ -1,5 +1,7 @@
 /// <reference types="../../scripts/types.d.ts" />
 
+declare module 'picomatch';
+
 declare interface HTMLInspectElement extends HTMLElement {}
 
 declare interface HTMLToggleElement extends HTMLElement {}
