+++ packages/client/src/elements/HTMLTreeElement/index.css
@@ -95,6 +95,7 @@
 }
 .oe-node {
   margin: 2px 0;
+  width: fit-content;
   font-size: 13px;
   font-weight: 500;
   color: var(--text-color);
