+++ packages/client/src/elements/utils/disableHoverCSS.ts
@@ -1,5 +1,4 @@
 import { createFrameChecker } from '../../utils/createFrameChecker';
-import { createAsyncTask } from '../../utils/createAsyncTask';
 
 const DISABLE_RE = /:hover/g;
 const DISABLE_TOKEN = '::hover';
@@ -66,4 +65,22 @@ function visitCSS(visitor: (css: string) => string) {
 function replaceRule(sheet: CSSStyleSheet, text: string) {
   sheet.deleteRule(0);
   sheet.insertRule(text, sheet.cssRules.length);
+}
+
+interface AsyncTask<T> extends Promise<T> {
+  resolve: (value: T | PromiseLike<T>) => void;
+  reject: (reason?: any) => void;
+}
+
+function createAsyncTask<T>() {
+  let resolve!: (value: T | PromiseLike<T>) => void;
+  let reject!: (reason?: any) => void;
+  const asyncTask = new Promise<T>((_resolve, _reject) => {
+    resolve = _resolve;
+    reject = _reject;
+  }) as AsyncTask<T>;
+  asyncTask.resolve = resolve;
+  asyncTask.reject = reject;
+
+  return asyncTask;
 }
