+++ packages/client/src/elements/utils/getColorMode.ts
@@ -1,4 +1,4 @@
-import { getOptions } from '../options';
+import { getOptions } from '../../options';
 
 const LightColorsCSS = css`
   :host {
