+++ packages/client/src/event/quickexit.ts
@@ -1,10 +1,10 @@
 import { omit } from '@open-editor/shared';
+import { off, on } from './index';
 import {
   type SetupListenerListener,
   type SetupListenerListenerOptions,
   createCustomEventHandler,
 } from './create';
-import { off, on } from '.';
 
 export default createCustomEventHandler('quickexit', setupListener);
