+++ packages/client/src/resolve/resolves/vue2.ts
@@ -1,6 +1,6 @@
-import type { SourceCodeMeta } from '../';
-import type { ResolveDebug } from '../resolveDebug';
 import { createVueResolver } from '../creators/createVueResolver';
+import type { SourceCodeMeta } from '../index';
+import type { ResolveDebug } from '../resolveDebug';
 
 let resolver: ReturnType<typeof createVueResolver<any>>;
 export function resolveVue2(
