+++ packages/client/src/resolve/resolves/vue3.ts
@@ -1,7 +1,7 @@
 import type { ComponentInternalInstance } from '@vue/runtime-core';
-import type { SourceCodeMeta } from '../';
-import type { ResolveDebug } from '../resolveDebug';
 import { createVueResolver } from '../creators/createVueResolver';
+import type { SourceCodeMeta } from '../index';
+import type { ResolveDebug } from '../resolveDebug';
 
 let resolver: ReturnType<typeof createVueResolver<ComponentInternalInstance>>;
 export function resolveVue3(
