+++ packages/client/src/setupClient.tsx
@@ -1,7 +1,7 @@
-import { appendChild } from './utils/ui';
-import { defineElements } from './elements';
+import { appendChild } from './utils/dom';
 import { InternalElements } from './constants';
 import { type Options, setOptions } from './options';
+import { defineElements } from './elements';
 
 export function setupClient(userOpts: Options) {
   if (!document.querySelector(InternalElements.HTML_INSPECT_ELEMENT)) {
