--- packages/client/src/utils/ui/index.ts
@@ -1,2 +1,0 @@
-export * from './dom';
-export * from './style';
